requirements.txt
django==4.1.0
djangorestframework==3.14.0
django-storages==1.13.1
pillow==9.3.0
boto3==1.26.14
django-rest-swagger==2.2.0
pyyaml==6.0
uritemplate==4.1.1